cast to double

+++ map_guides.cpp
@@ -626,7 +626,7 @@ main(const int argc, const char **argv) {
     
     if(VERBOSE){
       cerr << "# matches found = " << n_matches << endl;
-      cerr << "percent mapped = " << n_matches/sequences.size() << endl;
+      cerr << "percent mapped = " << static_cast<double>(n_matches)/sequences.size() << endl;
     }
     write_guides(counts_file_name, guides);
